Move opportunity status form to modules
