+++ src/apps/investments/client/opportunities/Details/state.js
@@ -8,8 +8,6 @@ export const TASK_SAVE_OPPORTUNITY_DETAILS = 'TASK_SAVE_OPPORTUNITY_DETAILS'
 export const TASK_SAVE_OPPORTUNITY_REQUIREMENTS =
   'TASK_SAVE_OPPORTUNITY_REQUIREMENTS'
 
-export const TASK_SAVE_OPPORTUNITY_STATUS = 'TASK_SAVE_OPPORTUNITY_STATUS'
-
 export const ID = 'opportunityDetails'
 
 export const state2props = (state) => state[ID]
