+++ src/client/modules/Investments/Opportunities/state.js
@@ -1,4 +1,5 @@
 export const TASK_CREATE_INVESTMENT_OPPORTUNITY =
   'TASK_CREATE_INVESTMENT_OPPORTUNITY'
+export const TASK_SAVE_OPPORTUNITY_STATUS = 'TASK_SAVE_OPPORTUNITY_STATUS'
 
 export const ID = 'opportunityDetails'
