+++ src/client/modules/Investments/Opportunities/tasks.js
@@ -5,3 +5,9 @@ export function createOpportunity(name) {
     .post('v4/large-capital-opportunity', name)
     .then(({ data: { id } }) => id)
 }
+
+export function saveOpportunityStatus({ opportunityId, values }) {
+  return apiProxyAxios.patch(`v4/large-capital-opportunity/${opportunityId}`, {
+    status: values.status,
+  })
+}
